Commit: - Optimized funcs of UserManager - Fixed refatch chats - Added some string - Added DeckTransition for ViewController - Realized ETH func (Create, Delete, Backup, Restore, TxHistory, Transfer, Show and share public key)

+++ paymon/Models/Money/CellCreatedMoneyData.swift
@@ -10,8 +10,8 @@ import Foundation
 
 class CellCreatedMoneyData : CellMoneyData {
     var fiatHint : String!
-    var currancyAmount : Decimal!
-    var fiatAmount : Decimal!
+    var currancyAmount : Double!
+    var fiatAmount : Double!
     var cryptoHint : String!
     var fiatColor : UIColor!
 }
